Accounts Controller Refactored Previously, the accoutns controller, in many cases, was not properly utilizing the ServiceError returned from the logic (services) layer. To correct this, the result returned from the logic layer is checked for a ServiceError before the controller returns anything to the user. If the result is a ServiceError, a failure response is sent to the user. Otherwise a success payload is sent.
Additionally, there was no message being passed to the failure methods throughout the
controller. Custom messages were passed for each message.

Lastly, some of the methods were renamed to be more implicit. The module name is accounts
so the method names can indicate only what it's doing, rather than what it's doing and
for whom it's doing it for.