BUG FIX
When querying certain routes in the accounts controller, the data
format being returned was incorrect. Instead of the data object containing
key value pairs of the accountDTO, it was a property named row with
an array of strings as the value.

By removing the brackets, now the data object is returned with the
proper format.