ENUM support added to accounts model.
In order to support the role system around a user, an ENUM was
used to define which values are acceptable as a role.

A user, when created, is REGULAR by default. Regular allows you
to view other user's content, but it does not allow you to
perform CRUD operations at all.

Right now, the only user with ADMIN privledges is myself. This
will allow me to grant SUPER privledges to others, as well as
perform CRUD operations on all users content.

When a user is a SUPER user, this will allow them to perform
CRUD operations on their own content, but this will not allow
them to peform CRUD operations on any other user's content.