BUG FIX
Originally, if you tried to update an existing user with an email or
username that another user has, there would have been a 500 server
error returned.

There was logic written into the service layer so that, before calling
the repository layer to update the user by that ID, there is a check to
see if a user already exists by that username or email. If either are
true, than there is a respective service error returned.