BUG FIX
The service layer returns a ServiceError if there is already a
user with the username or email that they are trying to update
with. Previously, if this operation was attempted, the server
would return a 500 response.

Now, that the service error is returned, there is logic written
for these cases and the respective HTTP status code is returned.