BUG FIXES
Firstly, in the create function, there was a syntax error with the
return statement being written as 'return' instead of 'returning'.

Secondly, the fetch_one function's call to the database did not
have the await syntax.

Lastly, there was a comma missing from the SET statement in the update
function.