BUG FIX / IMPROVEMENT
In the create_post method, when I was attempting to make an OR
comparison, I was using it with a bitwise operator, and not actually
utilizing Python's OR comparitor.

In addition, there was a problem with update_by_id, which would
force the user to provide values to the blog post and blog title.
To fix this, they were provided default values of None and there
was a check placed on the values being none.