BUG FIX
There was an issue with fetch_many blogs where the user was forced
to provide an account_id to make the request.

To fix this, the URI definition for this method was wrong for including
the account_id in the URI, which for a GET request, wouldn't work,
but also made the account_id manadtory for the request to work.