Bug Found / Improvements
fetch_by_email moved into it's own method. It was far too sloppy before,
having it combined with fetch many.

Service layer checks for database error and returns a service error based on that.
This is completely subject to chage.

In the accounts API, fetch_by_email there is a bug when you return a 204
if no email is found. The terminal returns an error regarding the length
of a header.